pyproject.toml
[project]
name = "adllm"
version = "0.1.0"
description = "Package for video ADL tagging using multimodal large language models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.1.1",
    "bitsandbytes>=0.44.1",
    "huggingface-hub>=0.26.2",
    "matplotlib>=3.9.2",
    "numpy>=2.1.3",
    "ollama>=0.4.0",
    "opencv-python>=4.10.0.84",
    "peft>=0.13.2",
    "pillow>=11.0.0",
    "polars>=1.14.0",
    "requests>=2.32.3",
    "safetensors>=0.4.5",
    "scikit-learn>=1.5.2",
    "seaborn>=0.13.2",
    "streamlit>=1.40.1",
    "torch>=2.5.1",
    "transformers>=4.46.3",
    "wordcloud>=1.9.4",
]
